Implemented basic searching state for guard enemy (#107)
Enemy transitions to searching if the player leaves the FOV cone while being chased. Simple implementation where enemy goes to the last known location of the player and looks left and right. Transitions to alerted in player is not found after completing search. Transitions to chasing again if player is spotted. Closes #81