#78 Implement standard guard enemy (#108)
* Added GuardPatrolState & refactored enemy entity framework

Added a guard patrol state for an enemy that follows a set path of waypoints continuously.

Refactored the enemy entity framework to better handle waypoints.

Added functionality to visualise enemy path using gizmos

Includes the following features:
- Patrols a set path of waypoints using NavMesh Plus for pathfinding
- Has field of view cone to show what is visible to enemy
- Transitions to alerted state when player enters field of view

* Add documentation

* Implemented basic alerted state for guard enemy (#103)

Enemy transitions to alerted when player enters it's FOV cone. Enemy continues on it's path but looks towards the player. Transitions to patrolling when player leaves the FOV cone. Transitions to chasing once player crosses threshold

* Implement basic chasing state for guard enemy (#104)

Implement ability for enemy to chase player once alerted to their position. Transition to attacking if player is in attack range. Transition to searching if player leaves FOV cone

* Implemented basic attack state for guard enemy and template for attack type (#106)

Enemy transitions to attacking when player is within it's attack range. Added a template for a shooting attack type. Transitions to chasing if player leaves the attack range and to searching is player leaves the FOV cone.

* Implemented basic searching state for guard enemy (#107)

Enemy transitions to searching if the player leaves the FOV cone while being chased. Simple implementation where enemy goes to the last known location of the player and looks left and right. Transitions to alerted in player is not found after completing search. Transitions to chasing again if player is spotted. Closes #81

* Documentation fixes

* Small bug fix

Fix a bug where the FOV cone for enemies does not display properly

* Small bug fix

Fixes a minor bug where the FOV cone does not display correctly when entering the searching state

* Small bug fix

Fixes a small bug where the FOV cone looks in the wrong direction when searching